+++ src/core/autocmds.ts
@@ -1,72 +1,126 @@
 'use strict'
 import { Neovim } from '@chemzqm/neovim'
+import { createLogger } from '../logger'
 import { Autocmd } from '../types'
 import { isFalsyOrEmpty } from '../util/array'
-import { crypto } from '../util/node'
+import { parseExtensionName } from '../util/extensionRegistry'
+import { omit } from '../util/lodash'
 import { Disposable } from '../util/protocol'
+const logger = createLogger('autocmds')
 
 interface PartialEnv {
   isVim: boolean
   version: string
 }
 
-const groupPrefix = 'coc_dynamic_'
+interface AutocmdOption {
+  group?: string | number
+  pattern?: string | string[]
+  buffer?: number
+  desc?: string
+  command?: string
+  once?: boolean
+  nested?: boolean
+  replace?: boolean
+}
 
-export function getAutoCmdText(autocmd: Autocmd): string {
-  let { arglist, pattern, request, callback } = autocmd
-  let res = ''
-  res += Array.isArray(autocmd.event) ? autocmd.event.join(' ') + ' ' : autocmd.event + ' '
-  if (pattern) res += pattern + ' '
-  if (request) res += 'request '
-  if (Array.isArray(arglist)) res += arglist.join(' ') + ' '
-  res += callback.toString()
-  return res
+export interface AutocmdOptionWithStack extends Autocmd {
+  stack: string
+}
+
+export class AutocmdItem {
+  private _extensiionName: string | undefined
+  constructor(
+    public readonly id: number,
+    public readonly option: AutocmdOptionWithStack
+  ) {
+  }
+
+  public get extensiionName(): string {
+    if (this._extensiionName) return this._extensiionName
+    this._extensiionName = parseExtensionName(this.option.stack)
+    return this._extensiionName
+  }
+}
+
+const groupName = 'coc_dynamic_autocmd'
+
+export function toAutocmdOption(item: AutocmdItem): AutocmdOption {
+  let { id, option } = item
+  let opt: AutocmdOption = { group: groupName }
+  if (option.buffer) opt.buffer = option.buffer
+  if (option.pattern) opt.pattern = option.pattern
+  if (option.once) opt.once = true
+  if (option.nested) opt.nested = true
+  let method = option.request ? 'request' : 'notify'
+  let args = isFalsyOrEmpty(option.arglist) ? '' : ', ' + option.arglist.join(', ')
+  let command = `call coc#rpc#${method}('doAutocmd', [${id}${args}])`
+  opt.command = command
+  return opt
 }
 
 export default class Autocmds implements Disposable {
-  public readonly autocmds: Map<string, Autocmd> = new Map()
+  public readonly autocmds: Map<number, AutocmdItem> = new Map()
   private nvim: Neovim
   private env: PartialEnv
+  private id = 0
 
   public attach(nvim: Neovim, env: PartialEnv): void {
     this.nvim = nvim
     this.env = env
   }
 
-  // unique id for autocmd to create unique group name
-  public generateId(autocmd: Autocmd): string {
-    let text = getAutoCmdText(autocmd)
-    return groupPrefix + crypto.createHash('md5').update(text).digest('hex')
-  }
-
-  public async doAutocmd(id: string, args: any[]): Promise<void> {
+  public async doAutocmd(id: number, args: any[]): Promise<void> {
     let autocmd = this.autocmds.get(id)
-    if (autocmd) await Promise.resolve(autocmd.callback.apply(autocmd.thisArg, args))
+    if (autocmd) {
+      let option = autocmd.option
+      // TODO add timeout limit for request
+      // autocmd.option.request
+      logger.trace(`Invoke autocmd from "${autocmd.extensiionName}"`, option)
+      try {
+        await Promise.resolve(option.callback.apply(option.thisArg, args))
+      } catch (e) {
+        e['stack'] = autocmd.option.stack
+        logger.error(`Error on autocmd "${option.event}"`, omit(option, ['callback', 'stack']), e)
+      }
+    }
   }
 
-  public registerAutocmd(autocmd: Autocmd): Disposable {
+  public registerAutocmd(autocmd: AutocmdOptionWithStack): Disposable {
     // Used as group name as well
-    let id = this.generateId(autocmd)
-    let { nvim } = this
-    this.autocmds.set(id, autocmd)
-    nvim.pauseNotification()
-    let cmd = createCommand(id, autocmd)
-    nvim.command('augroup ' + id, true)
-    nvim.command(`autocmd!`, true)
-    nvim.command(cmd, true)
-    nvim.command('augroup END', true)
-    nvim.resumeNotification(false, true)
+    let id = ++this.id
+    let item = new AutocmdItem(id, autocmd)
+    this.autocmds.set(id, item)
+    this.createAutocmd(item)
     return Disposable.create(() => {
-      nvim.command(`autocmd! ${id}`, true)
+      // only remove the item from autocmds
+      this.autocmds.delete(id)
     })
   }
 
-  public resetDynamicAutocmd(): void {
-    let { nvim } = this
+  private createAutocmd(item: AutocmdItem): void {
+    let { option } = item
+    let event = Array.isArray(option.event) ? option.event.join(',') : option.event
+    if (/\buser\b/i.test(event)) {
+      let cmd = createCommand(item.id, event, option)
+      this.nvim.command(cmd, true)
+    } else {
+      let opt = toAutocmdOption(item)
+      this.nvim.createAutocmd(item.option.event, opt, true)
+    }
+  }
+
+  public removeExtensionAutocmds(extensiionName: string): void {
+    let { nvim, autocmds } = this
     nvim.pauseNotification()
-    for (let [id, autocmd] of this.autocmds.entries()) {
-      nvim.command(`autocmd! ${id}`, true)
-      nvim.command(createCommand(id, autocmd), true)
+    nvim.command(`autocmd! ${groupName}`, true)
+    let items = autocmds.values()
+    for (const item of items) {
+      if (item.extensiionName === extensiionName) {
+        autocmds.delete(item.id)
+        continue
+      }
+      this.createAutocmd(item)
     }
     nvim.resumeNotification(false, true)
   }
@@ -76,13 +130,14 @@ export default class Autocmds implements Disposable {
   }
 }
 
-export function createCommand(id: string, autocmd: Autocmd): string {
+/**
+ * Only used for user autocmd, which can't be used for nvim_create_autocmd
+ */
+export function createCommand(id: number, event: string, autocmd: Autocmd): string {
   let args = isFalsyOrEmpty(autocmd.arglist) ? '' : ', ' + autocmd.arglist.join(', ')
-  let event = Array.isArray(autocmd.event) ? autocmd.event.join(',') : autocmd.event
-  let pattern = autocmd.pattern != null ? autocmd.pattern : '*'
-  if (/\buser\b/i.test(event)) {
-    pattern = ''
-  }
   let method = autocmd.request ? 'request' : 'notify'
-  return `autocmd ${id} ${event} ${pattern} call coc#rpc#${method}('doAutocmd', ['${id}'${args}])`
+  let opt = ''
+  if (autocmd.once) opt += ' ++once'
+  if (autocmd.nested) opt += ' ++nested'
+  return `autocmd ${groupName} ${event}${opt}  call coc#rpc#${method}('doAutocmd', [${id}${args}])`
 }