+++ src/handler/workspace.ts
@@ -38,7 +38,9 @@ export default class WorkspaceHandler {
     Object.defineProperty(window, 'openLocalConfig', {
       get: () => this.openLocalConfig.bind(this)
     })
-
+    extensions.onDidUnloadExtension(name => {
+      workspace.autocmds.removeExtensionAutocmds(name)
+    })
     commands.register({
       id: 'workspace.openLocation',
       execute: async (winid: number, loc: Location, openCommand?: string) => {
@@ -218,7 +220,7 @@ export default class WorkspaceHandler {
     })
   }
 
-  public async doAutocmd(id: string, args: any[]): Promise<void> {
+  public async doAutocmd(id: number, args: any[]): Promise<void> {
     await workspace.autocmds.doAutocmd(id, args)
   }